+++ lib/vault_db/core/api/api.dart
@@ -1,0 +1,6 @@
+export 'items/vault_items_api.dart';
+export 'system/categories_api.dart';
+export 'system/icons_api.dart';
+export 'system/tags_api.dart';
+export 'system/vault_system_api.dart';
+export 'vault_core_api.dart';