+++ lib/vault_db/providers/providers.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'api_providers.dart';
 export 'db_history_provider.dart';
 export 'decrypted_files_guard_provider.dart';
 export 'main_store_backup_orchestrator_provider.dart';